Restructure packages into "main" and "ruter"
As far as I understand it, every "thing" (Go package, repository etc.) should
have a "main" package. I'm mostly going by example[1], where all the files at
the root are in "main" package, and all files in sub directories have their own
package.

It also makes it convenient when developing locally, as `go run main.go` expects
the entry point to be main:main (`main` function in `main` package).

`RequestArrivalData` is renamed with a capital R, as that makes it a public
function apparently[2]:

 > In Go if something starts with a capital letter that means other packages
 > (and programs) are able to see it

In reality, I just did this because I didn't realise soon enough that you had to
pass multiple files to `go run` to make it compile correctly with all the
functions exported.

[1]:https://github.com/hashicorp/terraform
[2]:http://www.golang-book.com/11/index.htm#section1